Restore allowing signed int.

+++ src/AFQMC/Walkers/WalkerIO.hpp
@@ -248,13 +248,13 @@ bool restartFromHDF5(WalkerSet& wset, int nW_per_tg, std::string hdf_read_restar
     while(nread < nw_local) {
       if(ni+wlk_per_blk[bi] > nW0) {
         // determine block of walkers to read
-        size_t w0 = std::max(0,nW0-ni);
-        size_t nw_ = std::min(ni+wlk_per_blk[bi],nWN) - std::max(ni,nW0);
+        int w0 = std::max(0,nW0-ni);
+        int nw_ = std::min(ni+wlk_per_blk[bi],nWN) - std::max(ni,nW0);
         Data.reextent({nw_,wlk_nterms});
         hyperslab_proxy<boost::multi::array_ref<ComplexType,2>,2> hslab(Data,
-                                  std::array<size_t, 2>{static_cast<size_t>(wlk_per_blk[bi]), static_cast<size_t>(wlk_nterms)},
-                                  std::array<size_t, 2>{nw_,static_cast<size_t>(wlk_nterms)},
-                                  std::array<size_t, 2>{w0,0});
+                                  std::array<int, 2>{wlk_per_blk[bi], wlk_nterms},
+                                  std::array<int, 2>{nw_,wlk_nterms},
+                                  std::array<int, 2>{w0,0});
         read.read(hslab,std::string("walkers_")+std::to_string(bi));
         for(int n=0; n<nw_; n++, nread++)
           wset.copyFromIO(Data[n],nread);
